Fix -fwalk and -ffollow-jumps bug
I can't properly explain it and neither I can provide a concise test
case, but the bug happened when I tried to decompile the binary of SEGA
MegaDrive/Genesis game called DUNE 2 with reduced trace table. It is
huge (1 MiB) to take it as a test case and also probably illegal. So
anyway here is a fix.